+++ src/SafeRegex/Exception/CatastrophicBacktrackingException.php
@@ -1,0 +1,6 @@
+<?php
+namespace TRegx\SafeRegex\Exception;
+
+class CatastrophicBacktrackingException extends RuntimePregException implements PatternStructureException
+{
+}