+++ src/SafeRegex/Exception/JitStackLimitException.php
@@ -1,0 +1,6 @@
+<?php
+namespace TRegx\SafeRegex\Exception;
+
+class JitStackLimitException extends RuntimePregException implements PatternStructureException
+{
+}