+++ src/SafeRegex/Exception/JitStackLimitPregException.php
@@ -1,6 +1,0 @@
-<?php
-namespace TRegx\SafeRegex\Exception;
-
-class JitStackLimitPregException extends RuntimePregException implements PatternStructureException
-{
-}