+++ src/SafeRegex/Exception/RecursionException.php
@@ -1,0 +1,6 @@
+<?php
+namespace TRegx\SafeRegex\Exception;
+
+class RecursionException extends RuntimePregException implements PatternStructureException
+{
+}