+++ src/SafeRegex/Exception/RecursionLimitPregException.php
@@ -1,6 +1,0 @@
-<?php
-namespace TRegx\SafeRegex\Exception;
-
-class RecursionLimitPregException extends RuntimePregException implements PatternStructureException
-{
-}