+++ src/SafeRegex/Exception/SubjectEncodingException.php
@@ -1,0 +1,6 @@
+<?php
+namespace TRegx\SafeRegex\Exception;
+
+class SubjectEncodingException extends RuntimePregException
+{
+}