+++ src/SafeRegex/Exception/SubjectEncodingPregException.php
@@ -1,6 +1,0 @@
-<?php
-namespace TRegx\SafeRegex\Exception;
-
-class SubjectEncodingPregException extends RuntimePregException
-{
-}